Add a global logging abstraction to the govmm project
The project currently has a logging abstraction which is generic
in everything but name and location.  The interface is called
QMPLog and is stored in the qemu package.  This commit introduces
a new logging interface in its own package, vmmlog.  The
intention is that this logging interface will be used by govmm
packages and not just one.  It will soon replace the qemu.Logger
interface.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>