qemu: Fix the issue of missing reap qemu process
Some programs may have their own children reapers which has reaped the
qemu's process before cmd.Wait to reap it. For this case, the cmd.Run/cmd.Wait
will get an error contains substring of "no child processes", thus it's needed
to ignore this error and deal it as success.

Fixes: #65

Signed-off-by: fupan <lifupan@gmail.com>